[FIR] Disable assignment and `set` resolution for parenthesized LHS
This change implements the prohibition by disabling resolution to the
set operator altogether. While this sounds handy for performance
reasons, it leads to some errors being replaced. It's not a big deal,
since its red->red.

`CANNOT_INFER_PARAMETER_TYPE` appeared in `testListPlusAssign`, because
previously the call would have been resolved to `plus()`, but we are now
left with the unsuccessful `plusAssign()`. For the same reason some
`VAL_REASSIGNMENT` and `VARIABLE_EXPECTED` may disappear and be replaced
with the deprecation warning (see `LValueAssignment.latestLV.kt`).

^KT-70507 Fixed