[DLC] KtLightClassForDecompiledDeclaration: add missed kotlinOrigin to inner classes
Due to the bug, only inner classes with the same name as outer class had
`kotlinOrigin`.
Changes in LightClassUtil and KotlinDeclarationInCompiledFileSearcher are
covered by FirUastResolveApiFixtureTest#testResolvePropertiesInCompanionObjectFromBinaryDependency
test.
Effectively, this commit has a proper fix for KTIJ-27684
(which introduced additional logic in `LightClassUtil.getLightClassBackingField`,
so now we can safely drop it)

^KT-71469 Fixed