K2: forbid Array<Nothing?> usages in JVM platform
There is nothing bad in Array<Nothing?> in pure Kotlin.
Unfortunately, JVM has no good representation for this type:
in Kotlin, Array<Nothing?> is usable as Array<SomeType?>,
but in JVM, there is no type that is usable as SomeType[] when
SomeType can be arbitrary due to the nature of JVM arrays and
lack of the common lower bound in Java/JVM type system.
So here we have to forbid this type usages in JVM platform,
otherwise runtime exceptions as in KT-71528 are possible.

#KT-71528 Fixed