IR Inliner. Replace configuration key that switches double-inlining
Instead of the former `KlibConfigurationKeys.DOUBLE_INLINING_ENABLED`
use new `KlibConfigurationKeys.NO_DOUBLE_INLINING`, which is the
negated version of `DOUBLE_INLINING_ENABLED`.

The reason for this is the default value returned from
`CompilerConfiguration.getBoolean()` if no configuration key was found:
In case of the `CompilerConfiguration.getBoolean(DOUBLE_INLINING_ENABLED)`
call the default value would be `false` which would mean that
double-inlining is disabled. This is an undesired behavior, because even
in the absence of `DOUBLE_INLINING_ENABLED` key the default value
should be `true`.

^KT-71657