IR Inliner. Unwrap implicit casts when inlining function expressions
When double-inlining is turned on, it's possible to get a situation
at the 1st phase of inlining when an implicit cast is inserted for
a lambda argument of a non-private inline function call.
This prevents the inliner to inline this argument later at the 2nd
phase of inlining.

Illustrative example:
```
  inline fun <T> foo(block: () -> T): T = block()
  private inline fun <R> bar(block: () -> R): R = foo(block)

  fun test() {
    bar<String> { return "OK" }

    // after 1st phase of inlining (pseudocode):
    RETURNABLE_BLOCK {
      ...
      CALL foo(
        BLOCK type=kotlin.Function0<kotlin.Any?> {
          IMPLICIT_CAST type=kotlin.Function0<kotlin.Any?> {
          // ^^^ here is the problem
            FUN_EXPR type=kotlin.Function0<kotlin.String> origin=LAMBDA {
              FUN LOCAL_FUNCTION_FOR_LAMBDA {
                ...
                return "OK"
              }
            }
          }
        }
      )
    }
  }
```

The fix would be just to unwrap value arguments from `IMPLICIT_CAST`s
before inlining.

^KT-71659 Fixed