[Wasm] Fix member generation for data classes with an array-type property
When generating hashCode()/toString() for a data class, the compiler
generates code that uses the contentToString()/contentHashCode()
functions of the array property. It selects the overload that matches
the nullability of the property. The overloads with a non-null receiver
for arrays are deprecated (hidden) and should not be used. Instead,
use the nullable overloads.